+++ js/config.js
@@ -80,5 +80,5 @@ TD.urlMode = __params.get('mode') || null;
 TD.urlDiff = __params.get('diff') || null;
 TD.urlSeed = __params.get('seed') || null;
 TD.urlAutostart = __params.has('autostart') || __params.get('start') === '1';
-TD.VERSION = '1.4.0';
+TD.VERSION = '1.4.1';
 })();